app/requirements.txt
django
djangorestframework
djangorestframework-api-key
gunicorn

psycopg2-binary
django-redis

requests
base58
tronpy